Refactor MNC viewer widgets
